requirements.txt
n     asgiref==3.11.1
Django==6.0.2
et_xmlfile==2.0.0
lxml==6.0.2
numpy==2.4.2
openpyxl==3.1.5
pandas==3.0.0
python-dateutil==2.9.0.post0
python-docx==1.2.0
six==1.17.0
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
